spike-37: move json rpc init to main
