+++ nest-app/src/app.service.ts
@@ -2,7 +2,7 @@ import { CACHE_MANAGER, Inject, Injectable } from "@nestjs/common";
 import { Cron } from "@nestjs/schedule";
 import { Cache } from "cache-manager";
 import { GetStatusResult } from "casper-js-sdk";
-import { jsonRpc } from "./blocks/blocks.service";
+import { jsonRpc } from "./main";
 
 @Injectable()
 export class AppService {
