+++ nest-app/src/main.ts
@@ -1,6 +1,8 @@
 import { ValidationPipe } from "@nestjs/common";
 import { NestFactory } from "@nestjs/core";
+import { CasperServiceByJsonRPC } from "casper-js-sdk";
 import { AppModule } from "./app.module";
+import { nodeManager } from "./node-manager";
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, { cors: true });
@@ -10,3 +12,7 @@ async function bootstrap() {
   await app.listen(4000);
 }
 bootstrap();
+
+export const jsonRpc = new CasperServiceByJsonRPC(
+  nodeManager.getActiveNode().url
+);
